refactor: change component file structure to structure by routes
