+++ src/components/landing/Techs/Techs.js
